requirements.txt
pytorch
torchvision
tensorboardX
timm==0.3.2
